feat(server): compress streamed responses (RFC 9110 §8.4.1)
`CompressionMiddleware` returned a streamed body untransformed:

    // A streamed body is not transformed here (no streaming compression yet, P6).
    guard response.stream == nil else { return response }

So SSE, chunked downloads, generated bodies and every static file above
`FileResponder`'s 1 MiB streaming threshold were served uncoded no matter what
the client offered — which is the *inverse* of the right allocation of effort,
since those are the largest bodies on the wire.

They are now coded by a `CompressingBodyWriter` interposed between the
producer and the engine's writer. Each chunk is coded on its way past and
forwarded; nothing accumulates, and the codec keeps only its own fixed window.
The engine's writer stays the backpressure point, unchanged.

## Framing

`Content-Length` is removed and the stream's own `contentLength` becomes nil.
Both are necessary. The h1 engine frames a stream chunked precisely when
`stream.contentLength == nil` — but when it does, it adds
`Transfer-Encoding: chunked` and does **not** strip a `Content-Length` the head
already carries. Leaving one there would emit both fields together, which is
the RFC 9112 §6.1 request-smuggling shape, on top of being a length that
describes the identity body rather than the coded one. Pinned end-to-end over
`FakeConnection`, header assertions and a dechunk-then-gunzip of the wire.

## Negotiation and Vary

`Vary: Accept-Encoding` is now set before the buffered/streamed branch rather
than after the streamed early-return. Previously a streamed response got no
`Vary` at all, because the stream check preceded negotiation entirely — so a
cache could serve a compressible representation to a client that had asked for
a different coding. Both paths now go through the same `negotiatedEncoder` and
the same `addVary`, so a streamed and a buffered representation of one resource
carry byte-identical `Vary`.

Both paths also share `isEligible`, which lost its size check: a streamed body
may not know its own length. Each path now applies `minimumSize` to the length
it actually has — the buffered body's count, or `stream.contentLength`. A
stream of *unknown* length is coded, because not knowing how big it is, is the
reason it is streaming.

## What a compressing writer does with a `FileRegion`

It streams it through the compressor, forfeiting `sendfile(2)`. Not a
preference — by the time `writeFile(_:)` runs, `Content-Encoding` is already on
the wire, so handing the descriptor to the kernel would put identity octets
under a header promising a coding. The writer has no choice left; the choice
belongs upstream while the head is still mutable, and it is made there.

That placement is also what keeps zero-copy where it pays. `FileResponder`'s
precompressed sidecars set `Content-Encoding` themselves, so `isEligible`
declines them and the h1 `sendfile(2)` path survives for exactly the large
static assets it exists for. A large file with no sidecar trades the kernel
copy for the wire bytes, and the media types where that is a bad trade are
excluded before the head is written. No new threshold was invented for this:
one would have to live in the middleware anyway, and inventing a second tuning
knob without a measurement this host can support is what the last three
reverted changes on this branch were about.

## Cancellation

The codec lives in a `ContentCodingSession`, released by a `defer` on every
exit from the producer. A client that disconnects mid-download frees the
encoder's allocation at a point the code names, not whenever the last reference
to the producer closure happens to go. The session exists at all because
`ContentEncoderStream` is deliberately not `Sendable` while `ResponseStream`'s
producer is `@Sendable`; a `Mutex` rather than an actor, because coding a chunk
is synchronous CPU work and an actor hop per chunk would buy nothing.

## Retention

Measured, deterministically: 16 MiB streamed in 64 KiB chunks, every chunk
distinct so the codec cannot find the whole body in its window. Largest coded
chunk reaching the wire stays within 4x one input chunk for both gzip and
Brotli, and the final flush is small (profiled separately: 12 KB for ZLIB and
4 KB for Brotli after 8 MB of input). Had the codec held the body, the wire
would have seen one enormous chunk. No throughput or latency number is claimed
— this host has run concurrent agents throughout and nothing timed on it is
decision-grade.

`CompressionMiddlewareTests.skipsStreamingResponse` asserted the old gap and is
now `compressesStreamingResponse`. No public API changed.

+++ Sources/Core/CCRC32/crc32.c
@@ -71,7 +71,7 @@ uint32_t ccrc32_slice1(const uint8_t *buf, size_t len) {
 // MARK: - zlib (correct polynomial; internally hardware-accelerated)
 
 uint32_t ccrc32_zlib(const uint8_t *buf, size_t len) {
-    return ccrc32_update(crc32(0L, Z_NULL, 0), buf, len);
+    return ccrc32_update((uint32_t)crc32(0L, Z_NULL, 0), buf, len);
 }
 
 uint32_t ccrc32_update(uint32_t crc, const uint8_t *buf, size_t len) {
